API caller/requirements.txt
python=3.7
yfinance
tkinter
matplotlib.pyplot
datetime
dateutil.relativedelta
numpy
sys